Route-Me: Added Pixel-->Latitude/Longitude and Latitude/Longitude-->Pixel translation functions to the RMMapContents class:
-(CGPoint)latLngToPixel:(CLLocationCoordinate2D)latlong;
-(CLLocationCoordinate2D)pixelToLatLng:(CGPoint)pixel;

Added double-tap detection code that tests and verifies the functionality of the latLngToPixel and pixelToLatLng methods.  The double-tap code can be found in the touchesEnded withEvent method in the RMMapView class.

I fixed a small math bug in the projectMercatorPoint() method in the RMMercatorToScreenProjection class where it would not properly calculate the correct pixel.x coordinate with different combinations of positive/negative mercator points.

There is only one known bug with these LatLng/Pixel translation methods.  The map does not reset the Origin.x mercator value when it loops so the conversion from latLongToPixel will return an incorrect pixel.x value if you have panned the map enough to loop the map.

git-svn-id: https://route-me.googlecode.com/svn/trunk@53 1b297cfb-d155-0410-8e73-0b86fed47e86